+++ foxes/algorithms/sequential/__init__.py
@@ -1,3 +1,4 @@
 from .sequential import Sequential
+from .models import SequentialIter
 
 from . import models