+++ foxes/output/__init__.py
@@ -4,7 +4,7 @@ Output tools and functions.
 from .output import Output
 from .farm_layout import FarmLayoutOutput
 from .farm_results_eval import FarmResultsEval
-from .flow_plots_2d import FlowPlots2D
+from .flow_plots_2d.flow_plots import FlowPlots2D
 from .rose_plot import RosePlotOutput, StatesRosePlotOutput
 from .results_writer import ResultsWriter
 from .state_turbine_map import StateTurbineMap