+++ foxes/output/flow_plots_2d/__init__.py
@@ -1,0 +1,1 @@
+from .flow_plots import FlowPlots2D